% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_customTags}
\alias{add_customTags}
\title{add tags to a patient list}
\usage{
add_customTags(
  custom_patientLists,
  chosen_patientList,
  patient_list,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{custom_patientLists}{the custom patient lists}

\item{chosen_patientList}{the names of the chosen patient lists}

\item{patient_list}{dataframe, must contain column internalID}

\item{screentag}{add fomantic/semantic tags}

\item{screentag_print}{add printable/copyable text-only tags}
}
\value{
dataframe, with added column screentag or screentag_print
}
\description{
add tags to a patient list
}
