% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{custom_dataatable}
\alias{custom_dataatable}
\alias{datatableServer}
\title{Custom module - server}
\usage{
datatableServer(input, output, session, dMCustom)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMCustom}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Custom module - server
}
