% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Custom module - server}
\usage{
datatableServer(id, dMCustom)
}
\arguments{
\item{id}{id}

\item{dMCustom}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Custom module - server
}
