% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{initialize_data_table}
\alias{initialize_data_table}
\title{Custom module - initialize database table}
\usage{
initialize_data_table()
}
\value{
list
}
\description{
Custom module - initialize database table
}
