% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{read_configuration_db}
\alias{read_configuration_db}
\title{read the configuration database}
\usage{
read_configuration_db(dMeasureCustom_obj)
}
\arguments{
\item{dMeasureCustom_obj}{R6 object}
}
\description{
read the configuration database
}
