% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_patientList}
\alias{remove_patientList}
\title{remove patient list from configuration database}
\usage{
remove_patientList(dMeasureCustom_obj, name)
}
\arguments{
\item{dMeasureCustom_obj}{R6 object}

\item{name}{name of patient list}
}
\value{
TRUE if removed successfully
}
\description{
remove patient list from configuration database
}
