% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{update_patientList}
\alias{update_patientList}
\title{update patient list in configuration database}
\usage{
update_patientList(
  dMeasureCustom_obj,
  name,
  filename,
  column_ID = "ID",
  column_Label = "Label",
  keepAllColumns = FALSE,
  ID = NULL
)
}
\arguments{
\item{dMeasureCustom_obj}{R6 object}

\item{name}{name of patient list}

\item{filename}{CSV (comma-separated-value) filename
expects at least two columns,
'ID' and 'Label'}

\item{column_ID}{name of ID column}

\item{column_Label}{name of label column}

\item{keepAllColumns}{if FALSE, trim to just 'ID' and 'Label' columns}

\item{ID}{ID of patient list. If not provided, will try to infer from \code{name}}
}
\value{
ID of patient list.
 stop errors generated for several different reasons of failure
}
\description{
update patient list in configuration database
}
