% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{write_patientList}
\alias{write_patientList}
\title{write patient list to configuration database}
\usage{
write_patientList(
  dMeasureCustom_obj,
  name,
  filename,
  column_ID = "ID",
  column_Label = "Label",
  keepAllColumns = FALSE
)
}
\arguments{
\item{dMeasureCustom_obj}{R6 object}

\item{name}{name of patient list}

\item{filename}{CSV (comma-separated-value) filename
expects at least two columns,
'ID' and 'Label'}

\item{column_ID}{name of ID column}

\item{column_Label}{name of label column}

\item{keepAllColumns}{if FALSE, trim to just 'ID' and 'Label' columns}
}
\value{
ID of created patient list
}
\description{
write patient list to configuration database
}
