% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userInterface.R
\name{dMeasureConfigurationTabPanel}
\alias{dMeasureConfigurationTabPanel}
\title{Custom module - configuration panel server}
\usage{
dMeasureConfigurationTabPanel(id, dMCustom)
}
\arguments{
\item{id}{as required by Shiny modules}

\item{dMCustom}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Custom module - configuration panel server
}
