% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{dMeasureConfigurationTabPanel}
\alias{dMeasureConfigurationTabPanel}
\title{Custom module - configuration panel server}
\usage{
dMeasureConfigurationTabPanel(input, output, session, dMCustom)
}
\arguments{
\item{input}{as required by Shiny modules}

\item{output}{as required by Shiny modules}

\item{session}{as required by Shiny modules}

\item{dMCustom}{dMeasureCustom R6 object}
}
\value{
none
}
\description{
Custom module - configuration panel server
}
