% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{dMeasureConfigurationTabPanelItem}
\alias{dMeasureConfigurationTabPanelItem}
\title{Custom module - configuration tabpanel item}
\usage{
dMeasureConfigurationTabPanelItem()
}
\value{
tabPanel
}
\description{
Custom module - configuration tabpanel item
}
