% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom.R
\name{dMeasureConfigurationTabPanelUI}
\alias{dMeasureConfigurationTabPanelUI}
\title{Custom module - configuration panel UI}
\usage{
dMeasureConfigurationTabPanelUI(id)
}
\arguments{
\item{id}{module ID}
}
\value{
shiny user interface element
}
\description{
Custom module - configuration panel UI
}
